Avoid Int32 overflow in ApproximateSignedRankTest stdev.
On 32-bit, (nz+1) and (2nz+1) stay Int32 and the product can
overflow before promoting when multiplied by float(nz). Force
Float64 for every factor.

+++ src/wilcoxon.jl
@@ -273,7 +273,9 @@ function ApproximateSignedRankTest(x::Vector, W::Float64, ranks::Vector{T}, sign
     nz = length(ranks) # num non-zeros
     mu = W - nz * (nz + 1)/4
     # avoid integer overflow
-    std = sqrt(float(nz) * (nz + 1) * (2 * nz + 1) / 24 - tie_adjustment / 48)
+    # Promote every factor to Float64 before multiplying so Int32 (i686) cannot overflow
+    # nz*(nz+1)*(2nz+1) for nz ≳ 1290.
+    std = sqrt(float(nz) * float(nz + 1) * float(2 * nz + 1) / 24 - tie_adjustment / 48)
     ApproximateSignedRankTest(x, W, ranks, signs, tie_adjustment, n, median, mu, std)
 end
 function ApproximateSignedRankTest(x::AbstractVector{T}) where {T<:Real}
